requirements.txt
aiopg==1.4.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
decorator==5.1.1
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.114.0
fastapi-cli==0.0.5
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.1.1
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
prompt_toolkit==3.0.47
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
pure_eval==0.2.3
pycparser==2.22
pydantic==1.10.18
pydantic-sqlalchemy==0.0.9
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
rich==13.8.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.54
stack-data==0.6.3
starlette==0.38.5
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1